requirements-dev.txt
# Development dependencies for mortgage calculator

# Code quality and formatting
black==23.7.0
flake8==6.0.0
flake8-docstrings==1.7.0
isort==5.12.0
mypy==1.5.1
pylint==2.17.5
pre-commit==3.3.2

# Testing frameworks
pytest==7.4.3
pytest-flask==1.2.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development tools
ipython==8.15.0
ipdb==0.13.13
flask-shell-ipython==1.4.0

# Documentation
sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Security scanning
bandit==1.7.5
safety==2.3.1

# Performance testing
locust==2.16.1

# Debugging and profiling
flask-debugtoolbar==0.13.1
memory-profiler==0.61.0

# API testing
httpx==0.24.1
requests-mock==1.11.0

# Database tools (for future use)
alembic==1.11.3
flask-migrate==4.0.5
